### 2024-05-03 c425434
UltrasoundSystem.kspaceFirstOrder() : hotfix - missing 'psig' when using single transducer.
