### 2024-05-03 29e0dcc
ChannelData.scale : accept array inputs.
