### 2024-05-03 #6688 @adamltyson
Hi @thorstone25, I'll take over the submission from now on.

Would you be able to recommend appropriate reviewers for this submission? You can use the [list of potential reviewers](https://reviewers.joss.theoj.org/reviewers) to help. Please provide a list, but don't tag them (with @) so they aren't notified unnecessarily.

Thanks!
