### 2024-05-09 #6688 @thorstone25
Hi @adamltyson !

Thanks for taking over! This is my first submission, so please bear with me if I take a bit to get oriented.

From the list, I think the following profiles seem like good candidates:
- [at-jonbmartin](https://github.com/jonbmartin)
- [at-mrava87](https://github.com/mrava87)
- [at-tomelse](https://github.com/tomelse)
- [at-rafaelorozco](https://github.com/rafaelorozco)
- [at-pankajkmishra](https://github.com/pankajkmishra)

Thanks!
