### 2024-05-17 #6688 @adamltyson
Thanks @jonbmartin, @rafaelorozco!

@tomelse, @ebezzam - would either of you also be up for reviewing this?
