### 2024-05-23 #6772 @editorialbot
[REVIEW]: QUPS: A MATLAB Toolbox for Rapid Prototyping of Ultrasound Beamforming and Imaging Techniques
**Submitting author:** @thorstone25 (<a href="http://orcid.org/0000-0002-0720-6955">Thurston Brevett</a>)
**Repository:** https://github.com/thorstone25/qups
**Branch with paper.md** (empty if default branch): dev-paper
**Version:** v1.2.1
