### 2024-05-24 362f935
UltrasoundSystem.focusTx() : set default block size to all Txs; cosmetic
