### 2024-05-24 20dd986
wsinterpd2() : CUDA/OpenCL kernel calls cached; avoid OpenCL file lookup if native CUDA branch selected.
