### 2024-06-04 #6772 @rafaelorozco
Hello, Everything looks great. I was able to install and run tests successfully as described in the repo.

I have only left **Community guidelines** unchecked since I couldnt find instructions on how to contribute to the software. After this addition, I highly recommend the publication of this manuscript to JOSS. Congratulations
@thorstone25 on the work for this package
