### 2024-06-04 #6772 @adamltyson
Thanks for the quick turnaround @rafaelorozco!

@thorstone25 could you add community guidelines to the repository so @rafaelorozco can tick off this section:
>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
