### 2024-06-06 9f9e882
Create CONTRIBUTING.md
