+++ CONTRIBUTING.md
@@ -3,7 +3,7 @@
 Not every unexpected result or error is a bug - before confirming that you've found a bug, check for the following gotchas:
 - [ ] Are the units correct? For example, are *all* units in meters / seconds / Hertz ? Hint: use `UltrasoundSystem.plot()` as a quick sanity check!
 - [ ] Do the data sizes correspond? For example, does `us.rx.numel == chd.N` and `us.seq.numPulse == chd.M` ? 
-- [ ] Is this an unimplemented feature? For example, the f-k migration beamformer expects a `TransducerArray` not a `TransducerMatrix`.
+- [ ] Is this an unimplemented feature? For example, the f-k Stolt's migration beamformer expects a `TransducerArray` not a `TransducerConvex`.
 - [ ] Is there a theoretical limitation? For example, a CFL > 0.3 may lead to instability or a sampling frequency < Nyquist may lead to aliasing.
 - [ ] Are you running out of memory? This ... is probably not a bug, but can usually be handled by a for-loop, or using a `bsize` (block size) input if available.
 - [ ] Is it a parallel pool incompatibility? This may be intentionally maintained for forward compatibility, as later MATLAB versions have [relaxed restrictions on parallel pools](https://www.mathworks.com/help/parallel-computing/release-notes.html).
@@ -18,7 +18,7 @@ If none of the above addresses the bug, please submit an issue, and include:
 ### Submitting a Patch
 If you also have a patch that fixes a bug, after submitting an issue, feel free to open a pull request.
 
-The patch does not need to follow all stylistic coding conventions, but it must pass all of the existing tests in QUPS. It is recommended but not required to add additional tests within an appropriate test class that addresses the bug.
+The patch does not need to follow all stylistic coding conventions, but any changes must pass all of the existing tests in QUPS. It is recommended but not required to add additional tests within an appropriate test class that addresses the bug.
 Please reference the issue when submitting a pull request.
 
 ### New Features
@@ -29,16 +29,12 @@ In any case, there is no gaurantee that a request will be implemented.
 If you have implemented a new feature and would like for it to be incorporated, feel free to open a pull request.
 For a new feature to be accepted, it should conform to the appropriate [class structure](src/README.md#class-structure) and [data format](src/README.md#data-format) conventions to maximize compatibility.
 
-### New Examples
-Example scripts are more than welcome!
-If there is a concept or method where an example script would be useful, especially as a starting template, please feel free to create a pull request.
-While this package does not emphasize research applications, these examples are also within scope.
-If required, examples should either create data from within QUPS (e.g. using `UltrasoundSystem.greens`) or reference a downloadable public dataset. Submissions which include data will be rejected.
-
-### How-to Requests
+### Examples & How-tos 
 QUPS is a toolbox, not a library: for most applications there is some assembly required. 
-It's assumed that you have sufficient domain knowledge to identify from the documentation which tools to use and how.
-If you are unsure about what to implement or which function to use, you will likely need to consult other resources for a more detailed explanation of ultrasound terminology, methods, and practices.
+Example scripts that clarify syntax, applications, principles, or concepts are more than welcome! 
+If you believe an example would be useful to the community, please feel free to create an issue to solicit an example, or a pull request to contribute one.
 
-Thanks!
+Example scripts should clearly demonstrate one or two ideas and complete in a few minutes at most.
+If required, contributed examples should either create data from within QUPS (e.g. using `UltrasoundSystem.greens()`) or reference a downloadable public dataset.
+Submissions which include data will be rejected.
 
