### 2024-06-09 cbec448
Sequence.Verasonics() : scalarize t0 when possible.
