### 2024-06-09 a9dd15b
ChannelData() : added mtimes() - matrix multiplication for ChannelData; update documentation.
