### 2024-06-09 ad538c5
ChannelData.data - default to 0x0x0 so chd.t == chd.N == chd.M == 0 on init.
