+++ src/Medium.m
@@ -340,7 +340,7 @@ classdef Medium < matlab.mixin.Copyable
 
     % fullwave interface
     methods
-        function maps = getFullwaveMap(med, scan)
+        function maps = getFullwaveMap(med, grd)
             % GETFULLWAVEMAP - Get Fullwave compatible map structure
             %
             % maps = getFullwaveMap(med, scan) returns a map sampled on
@@ -375,11 +375,11 @@ classdef Medium < matlab.mixin.Copyable
 
             arguments
                 med Medium
-                scan Scan
+                grd ScanCartesian
             end
 
             % sample all maps on the grid points
-            [c, rho, BoA, alpha] = props(med, scan);
+            [c, rho, BoA, alpha] = props(med, grd);
 
             % set the map properties
             eta = 1 + BoA./2;
@@ -394,7 +394,7 @@ classdef Medium < matlab.mixin.Copyable
 
     % k-Wave interface
     methods
-        function kmedium = getMediumKWave(med, scan)
+        function kmedium = getMediumKWave(med, grd, kwargs)
             % GETMEDIUMKWAVE - Get a kWave compatible medium struct
             %
             % kmedium = GETMEDIUMKWAVE(med, scan) creates a kWave
@@ -429,14 +429,14 @@ classdef Medium < matlab.mixin.Copyable
             % See also ULTRASOUNDSYSTEM.KSPACEFIRSTORDER MEDIUM.PROPS
             arguments
                 med Medium
-                scan Scan
+                grd ScanCartesian
             end
 
             % get properties in original dimensions
-            [c, rho, BoA, alpha] = med.props(scan);
+            [c, rho, BoA, alpha] = med.props(grd);
 
             % get k-Wave order
-            ord = arrayfun(@(d) find(d == scan.order), 'ZXY'); % place in this order for kWave
+            ord = arrayfun(@(d) find(d == grd.order), 'ZXY'); % place in this order for kWave
 
             % move to k-Wave dimensions
             [kmedium.sound_speed, kmedium.density, kmedium.BonA, kmedium.alpha_coeff] = ...
@@ -588,11 +588,11 @@ classdef Medium < matlab.mixin.Copyable
             medium.pertreg = [{@(p) dealret(p, cfun, rfun, bfun, afun)}, medium.pertreg];
         end
     
-        function med = Diffuse(us, grid, kwargs)
+        function med = Diffuse(us, grd, kwargs)
             % DIFFUSE - Generate a diffuse (density) scattering Medium
             %
-            % med = DIFFUSE(us, grid) generates a diffuse scaterring Medium
-            % med on the ScanCartesian grid for the UltrasoundSystem us.
+            % med = DIFFUSE(us, grd) generates a diffuse scaterring Medium
+            % med on the ScanCartesian grd for the UltrasoundSystem us.
             %
             % rho = DIFFUSE(..., 'scat_per_cell', n) sets the number of
             % scatterers per resolution voxel. A rule of thumb minimum is 12. The
@@ -613,22 +613,22 @@ classdef Medium < matlab.mixin.Copyable
             % med = Medium();
             %
             % % Get a simulation region
-            % grid = copy(us.scan);
-            % [grid.dx, grid.dy, grid.dz] = deal(us.lambda / 8); % set grid resolution
+            % grd = copy(us.scan);
+            % [grd.dx, grd.dy, grd.dz] = deal(us.lambda / 8); % set grid resolution
             %
             % % Create a diffuse scattering distribution
-            % med = Medium.Diffuse(us, grid);
+            % med = Medium.Diffuse(us, grd);
             %
             % % display
[…]
