### 2024-06-09 30edfde
Add test for chd * chd.
