### 2024-06-10 107bf9e
Add Contributing Guidelines (#35)
* Create CONTRIBUTING.md

* Update CONTRIBUTING.md

* Update CONTRIBUTING.md
