### 2024-06-10 #6772 @thorstone25
Thank you @rafaelorozco ! I appreciate you taking the time to review it!

@adamltyson and @rafaelorozco, I just added a CONTRIBUTING.md file to the main branch of the repository. Is that sufficient?
