### 2024-06-17 ec6bd3e
Medium.m : bug fix (commit artefact.
