### 2024-06-24 cf6e3a2
Scan*() : expand tolerance for evenly spaced axes.
