### 2024-06-24 e6c8817
pwznxcorr() : normalize to 1 when w is multi-channel.
