### 2024-06-30 5290aa4
pwznxcorr()  :  sizing  bug fix.
