### 2024-06-30 5649616
wsinterpd2() : avoid checking kernel cache if GPU unavailable.
