### 2024-07-01 67f00c0
UltrasoundSystem.bf* : beamformers support separable apodization functions; rename '*scan' to 'cgrd' for clarity.
