+++ kern/das_spec.m
@@ -91,7 +91,7 @@ function [y, k, pre_args, post_args] = das_spec(fun, Pi, Pr, Pv, Nv, x, t0, fs, 
 VS = true; % whither plane wave
 DV = false; % whither diverging wave
 interp_type = 'linear'; 
-apod = 1;
+apod = cell(1,0);
 isType = @(c,T) isa(c, T) || isa(c, 'gpuArray') && strcmp(classUnderlying(c), T);
 if isType(x, 'single')
     idataType = 'single';
@@ -131,9 +131,9 @@ while (n <= nargs)
         case 'interp'
             n = n + 1;
             interp_type = varargin{n};
-        case 'apod',
+        case 'apod'
             n = n + 1; 
-            apod = varargin{n};
+            apod{end+1} = varargin{n};
         case 'modulation'
             n = n + 1;
             fmod = varargin{n};
@@ -155,6 +155,9 @@ if (nargin < 8 || isempty(fs))
         error('Undefined sampling rate.');
     end
 end
+
+% ensure non-empty
+if isempty(apod), apod = {1}; end
 
 % init
 [k, pre_args, post_args] = deal({});
@@ -187,6 +190,11 @@ end
 % dispatch
 if device && (gdev || odev)
 
+    % ocl requires exactly this many separable apodization matrices
+    AMAX = 1; % can be modified by modifying the kernel
+    if ~gdev && numel(apod) > AMAX, error("QUPS:das_spec:OpenCLApodizationSupport","OpenCL only supports up to "+AMAX+" apodization matrices."); end
+    if ~gdev, [apod{end+1:AMAX}] = deal(1); end
+    
     % warn if non-linear interp was requested
     switch interp_type
         case "nearest", flagnum = 0;
@@ -218,7 +226,7 @@ if device && (gdev || odev)
         inps = {Pi, Pr, Pv, Nv, x, t0, fs, cinv};
         oclass = cellfun(@class, inps, per_cell{:}); 
         tmp = cellfun(@gather, inps, per_cell{:});
-        g = gpuDevice(device); 
+        % g = gpuDevice(device); 
         tmp = cellfun(@(inp, cls)cast(inp, cls), tmp, oclass, per_cell{:});
         [Pi, Pr, Pv, Nv, x, t0, fs, cinv] = deal(tmp{:});
     end
@@ -232,9 +240,9 @@ if device && (gdev || odev)
         if odev, ptypefun = @(x)          single(x) ; end
         dtypefun = @(x) complex(halfT(x));
     end
-    [x, apod] = dealfun(dtypefun, x, apod);
+    [x, apod{:}] = dealfun(dtypefun, x, apod{:});
     [Pi, Pr, Pv, Nv, cinv] = dealfun(ptypefun, Pi, Pr, Pv, Nv, cinv);
-    
+
     % expand all inputs to 3D
     expand_inputs();
     
@@ -242,12 +250,13 @@ if device && (gdev || odev)
     [T, N, M] = size(x, 1:3);
     Isz = size(Pi, 2:4); % I1 x I2 x I3 == I
     I = prod(Isz);
+    S = numel(apod);
 
     % get stride for apodization and sound speed
-    Iasz = size(apod,1:5);
-    Icsz = size(cinv,1:5);
-    astride = [1, cumprod(Iasz(1:end-1))] .* (Iasz ~= 1);
-    cstride = [1, cumprod(Icsz(1:end-1))] .* (Icsz ~= 1);
+    Icsz = size(cinv,1:5)';
+    Iasz = cell2mat(cellfun(@(a){size(a,1:5)'}, apod));
[…]

[…]
