### 2024-07-06 2895a19
UltrasoundSystem : commit artefact bug fix
