### 2024-07-07 d70d9b5
sub() : data type handles tall dim; variable name change for optimistic in-place operation.
