### 2024-07-12 e0098da
UltrasoundSystem.{DAS, getDASConstCUDADef}() : handles transposition of rx/tx without a data copy; const complie for num apodization, interpolation, transposition
