+++ src/UltrasoundSystem.m
@@ -2898,12 +2898,11 @@ classdef UltrasoundSystem < matlab.mixin.Copyable & matlab.mixin.CustomDisplay
                 % make sure t0 is a scalar in all dims except transmit
                 if ~all(size(chd.t0, setdiff(1:ndims(chd.t0), chd.mdim)) == 1), warning("Resampling data for a scalar t0."); chd = rectifyt0(chd); end
 
-                % data must be ordered T x N x M
-                ord = [chd.tdim, chd.ndim, chd.mdim];
-                if ~isequal(ord, 1:3), chd = rectifyDims(chd); end % reorder if necessary
+                % data must be ordered T x perm(N x M) x ...
+                if chd.tdim ~= 1, chd = rectifyDims(chd); end % reorder if necessary
 
                 % get the beamformer arguments
-                dat_args = {chd.data, gather(chd.t0), gather(chd.fs), c0, 'device', kwargs.device, 'input-precision', kwargs.prec}; % data args
+                dat_args = {chd.data, gather(chd.t0), gather(chd.fs), c0, 'device', kwargs.device, 'input-precision', kwargs.prec, 'transpose', (chd.ndim>chd.mdim)}; % data args
                 if isfield(kwargs, 'interp'), interp_args = {'interp', kwargs.interp}; else,  interp_args = {}; end
                 ext_args = [interp_args, apod_args]; % extra args
     
@@ -2924,10 +2923,10 @@ classdef UltrasoundSystem < matlab.mixin.Copyable & matlab.mixin.CustomDisplay
                 % request the CUDA kernel?
                 if nargout > 1, ext_ret = cell(1,3); else, ext_ret = {}; end
 
-                % beamform the data (I1 x I2 x I3 x N x M x F x ...)
+                % beamform the data (I1 x I2 x I3 x perm(N x M) x F x ...)
                 [b, ext_ret{:}] = das_spec(fun, pos_args{:}, dat_args{:}, ext_args{:});
 
-                % move data dimension, back down raise aperture dimensions (I1 x I2 x I3 x F x ... x N x M)
+                % move data dimension, back down raise aperture dimensions (I1 x I2 x I3 x F x ... x perm(N x M))
                 b = permute(b, [1:3,6:(D+2),4:5]);
 
                 % store
@@ -5015,7 +5014,7 @@ classdef UltrasoundSystem < matlab.mixin.Copyable & matlab.mixin.CustomDisplay
                 end
             end
         end
-        function def = getDASConstCudaDef(us, chd)
+        function def = getDASConstCudaDef(us, chd, varargin, kwargs)
             % GETDASCONSTCUDADEF - Constant size compilation definition for DAS
             %
             % def = GETDASCONSTCUDADEF(us) creates a compilation
@@ -5028,12 +5027,51 @@ classdef UltrasoundSystem < matlab.mixin.Copyable & matlab.mixin.CustomDisplay
             % at the cost of compile time and flexibility.
             %
             % After compiling with this definition, calls with a different
-            % size of data will have unexpected results. Use
-            % UltrasoundSystem.recompileCUDA to reset the binaries to
-            % handle variable sizes. 
+            % input or output data size will have unexpected results. Use
+            % us.recompileCUDA() to reset the binaries to handle variable sizes. 
             % 
             % def = GETDASCONSTCUDADEF(us, chd) additionally uses a fixed
             % size ChannelData object.
+            % 
+            % def = GETDASCONSTCUDADEF(us, chd, a1, a2, ..., an)
+            % additionally sets the number of apodization matrices. If not
+            % included, the default number of apodization matrices is 0.
+            % 
[…]

[…]
