### 2024-07-12 12f4b46
das_spec() : bug fix.
