### 2024-07-16 a2478f5
ScanPolar() : update help example.
