### 2024-07-16 6eed3af
ScanSpherical() : update defaults, implement ScanConvert(), bug fix.
