+++ src/ScanSpherical.m
@@ -7,7 +7,7 @@ classdef ScanSpherical < Scan
     properties
         r (1,:) {mustBeVector} = 1e-3*linspace(0,40,161)  % image range values
         a (1,:) {mustBeVector} = linspace(-45,45,91)      % image azimuth angle values (deg)
-        e (1,:) {mustBeVector} = 0                        % image elevation angle values (deg)
+        e (1,:) {mustBeVector} = linspace(-45,45,91)      % image elevation angle values (deg)
         order = 'RAE';                      % pixel order of change 
         % origin (3,1) = [0;0;0];             % center of the coordinate system with respect to cartesian coordinates (m)
     end
@@ -128,31 +128,63 @@ classdef ScanSpherical < Scan
             % b_cart = SCANCONVERT(self, b_pol, scanC) uses a given
             % ScanCartesian scanC instead of creating one.
             % 
+            % Example:
+            % % Create a system
+            % xdc = TransducerMatrix.PO192O();
+            % xdc.numd = [8 8]; % shrink to 8 x 8 
+            % scn = ScanSpherical('a',-25:25, 'e', -25:25, 'r', [0 40e-3]);
+            % us = UltrasoundSystem('xdc', xdc, 'scan', scn);
+            % scn.dr = us.lambda / 2; % set resolution
+            % 
+            % % Simulate echoes
+            % sct = Scatterers.Grid([3 2 4], 5e-3, [0 0 20e-3]', 'c0', us.seq.c0);
+            % chd = greens(us, sct);
+            % 
+            % % Beamform on the polar grid
+            % b = DAS(us, chd);
+            % 
+            % % Scan convert to the cartesian grid
+            % b = mod2db(b); % envelope-detection, log-compression
+            % [bc, scnc] = scanConvert(us.scan, b);
+            % 
+            % % Display
+            % figure; whitebg('k'); tiledlayout('flow');
+            % h1 = vol3d(scn , b , nexttile(), ceil( scn.size(3)/2)); 
+            % h2 = vol3d(scnc, bc, nexttile(), ceil(scnc.size(3)/2)); hold on; 
+            % hx = patch(xdc,'FaceColor',[1,1/2,0]); 
+            % hs = plot(sct, 'r.');
+            % legend([hx hs]);
+            % 
+            % % Format display
+            % title(h1.parent, "Spherical (Original)" );
+            % title(h2.parent, "Cartesian (Converted)");
+            % for ax = [h1.parent, h2.parent]
+            %     axes(ax); % make current (sets gca)
+            %     ax.YDir = 'reverse'; view(3); % set view
+            %     dbr b-mode 60; alim(max(caxis)+[-80 -0]); % set colors
+            % end
+            % 
             % See also SCANCARTESIAN
 
-            error("Not implemented.");
+            % create an output scan if one not given
+            if nargin < 3, scanC = ScanCartesian(self); end
 
-            % % create an output scan if one not given
-            % if nargin < 3, scanC = ScanCartesian(self); end
-            % 
-            % % get the cartesian points for the output scan
-            % [X, Y, Z] = scanC.getImagingGrid();
-            % 
-            % % convert to spherical
-            % og = self.origin;
-            % [A, E, R] = cart2sph(Z - og(3), X - og(1), Y - og(2));
-            % 
-            % % sample the data at these coordinates
-            % % TODO: handle different data orders / dimensionality via 
-            % % permution / page functions
-            % assert(self.order == "RAE", "Data must be in order 'RAE'.");
[…]
