### 2024-07-16 8375e15
dbr() : apply to image or surface in gca.
