+++ test/ExampleTest.m
@@ -60,7 +60,8 @@ classdef ExampleTest < matlab.unittest.TestCase
                 'gpu',            "wbilerpg", ... CUDAKernel or oclKernel support required
                 'comp',     ("recompile" + ["","Mex","CUDA"]) ... compilations setup required
                 }, 2, [])'; % + "(";
-            
+            bl_fcn(:,2) = cellfun(@(s){s+"("}, bl_fcn(:,2)); % append "("
+
             % reference or load Qups.prj if not loaded
             prj_rt = fullfile(mfilename("fullpath"),".."); % Qups directory
             try
@@ -92,7 +93,7 @@ classdef ExampleTest < matlab.unittest.TestCase
                 bl_fcn(bl_fcn(:,1) == "gpu",:) = []; 
             end
             kerns = ["msfm2d", "msfm3d"]+"."+mexext; % require mex binaries
-            if all(arrayfun(@(k) exist(fullfile(prj_rt, "bin", k), 'file'), kerns))
+            if 1||all(arrayfun(@(k) exist(fullfile(prj_rt, "bin", k), 'file'), kerns))
                 bl_fcn(bl_fcn(:,1) == "mex",:) = []; 
             end
             % fullwave executable must exist in 'bin' folder under this name
@@ -104,9 +105,9 @@ classdef ExampleTest < matlab.unittest.TestCase
 
             % filter by gpu compiler availability (system not available on thread pools)
             % (not quite working - seems the environment changes?)
-            % if ~isempty(argn(2, @system, "which nvcc"))
-                % bl_fcn(bl_fcn(:,1) == "comp",:) = [];
-            % end
+            if ~isempty(argn(2, @system, "which nvcc"))
+                bl_fcn(bl_fcn(:,1) == "comp",:) = [];
+            end
         end
         function bl_file = generateBlacklistFiles()
             bl_file = [ ...
@@ -119,10 +120,7 @@ classdef ExampleTest < matlab.unittest.TestCase
     methods
         function filterBlacklist(test, code, blk, fls)
             % check for blacklisted variables or functions
-            for i = length(test.bl_fcn(:,1)):-1:1
-                m(:,i) = contains(code, test.bl_fcn{i,2} + "("); % lines x pcks
-            end
-            m   = [m    , contains(code, test.bl_var)]; % lines x (pcks+1)
+            m = cell2mat(cellfun(@(s) {contains(code, s)}, [test.bl_fcn(:,2); {test.bl_var}]')); % lines x (pcks+1)
             pkg = [string(test.bl_fcn(:,1))', "var"]; % packages
             l = any(m,2); % matching lines
             p = any(m,1); % matching packages
@@ -130,7 +128,7 @@ classdef ExampleTest < matlab.unittest.TestCase
                 join(rmmissing([ ...
                 "[blacklist] Filtering lines " + join(string(blk([1 end])),"-") + " in file " + fls + "." ...
                 ,"Matching line(s):" + newline + join(code(l), newline) ...
-                ,"Matching packages: " + join(string(pkg(p)), ", ") ...
+                ,"Matching packages: " + join(string(pkg(1,p)), ", ") ...
                 ]), newline));
         end
         function silenceAcceptableWarnings(testCase)
