### 2024-07-16 0e8c4b0
wsinterpd2() : modify caching strategy.
