### 2024-07-16 9e40572
Update example, cheat_sheet scripts.
