+++ cheat_sheet.m
@@ -156,6 +156,17 @@ scan = ScanPolar('r',r,'a',a,'y',y);
 scan.dr = 0.1e-3;
 scan.da = 0.25;
 
+% setup a 3D spherical imaging region
+r =   0 : 0.2e-3 : 40e-3;
+a = -40 : 0.5    : 40   ;
+e = -40 : 0.5    : 40   ;
+scan = ScanSpherical('r',r,'a',a,'e',e);
+
+% set the resolution for a Spherical region
+scan.dr = 0.1e-3;
+scan.da = 0.25;
+scan.de = 0.25;
+
 % create a Generic Scan with an arbitrary axes-to-pixel transform function
 scan = ScanGeneric('u', 1:100, 'v', -90:90, 'w', 1 : 100);
 scan.trans = @(u,v,w) cat(1,u,cosd(v), exp(w));
@@ -163,9 +174,12 @@ scan.trans = @(u,v,w) cat(1,u,cosd(v), exp(w));
 %% Beamforming
 % create some example objects and data for this section
 us = UltrasoundSystem(); % default system
+seq = Sequence('type', 'FC', 'focus', [0 0 50e-3]' + [1 0 0]'*(-20:20));
 scat = Scatterers('pos', [0;0;30e-3], 'c0', us.seq.c0); % point scatterer(s)
 med = Medium('c0', scat.c0); % FDTD medium
 chd = greens(us, scat); % get some data for this example 
+usf = setfield(copy(us),'seq',seq); % focused
+chdf = us.focusTx(chd, seq); % focused data
 
 % ------------- B-mode Images ----------------- %
 % standard delay-and-sum (DAS)
@@ -185,10 +199,10 @@ b = bfEikonal(us, chd, med);
 
 % eikonal equation beamformer with a high sound speed resolution grid
 % currently lambda/10 or finer recommended (FSA only)
-cscan = copy(us.scan); 
-cscan.dx = us.lambda / 10; 
-cscan.dz = us.lambda / 10;
-b = bfEikonal(us, chd, med, cscan);
+grd = copy(us.scan); 
+grd.dx = us.lambda / 10; 
+grd.dz = us.lambda / 10;
+b = bfEikonal(us, chd, med, grd);
 
 % frequency-domain adjoint Green's function beamformer
 % Note: you may get poor performance on virtual source ('FC'/'DV'/'VS')
@@ -219,12 +233,18 @@ bd = dmas(bn, ndim);
 % 45 deg acceptance angle weights on receive
 a = us.apAcceptanceAngle(45); 
 
-% set aperture growth to limit f# >= 1
-a = us.apApertureGrowth(1);
+% set aperture growth to limit f# >= 1.5
+a = us.apApertureGrowth(1.5);
+
+% use multi-line beamformer weights (focused transmits)
+am = usf.apMultiline();
 
 % apply apodization when beamforming 
 % (works in most cases with most beamformers)
-b = DAS(us, chd, 'apod', a);
+b = DAS(us, chd, a);
+
+% apply multiple apodization schemes simultaneously
+b = DAS(usf, chdf, a, am);
 
 %% Channel Data
 % This section demonstrates operations that can be efficiently done on
@@ -240,6 +260,7 @@ b = DAS(us, chd, 'apod', a);
 % create some example objects and data for this section
 us = UltrasoundSystem(); % default system
 seq = SequenceRadial('type','PW','angles',-10:10); % alternate pulse sequence
+wv = us.xdc.xdcImpulse(); % waveform
 scat = Scatterers('pos', [0;0;30e-3], 'c0', us.seq.c0); % point scatterer(s)
[…]
