+++ test/KernTest.m
@@ -5,15 +5,37 @@ classdef KernTest < matlab.unittest.TestCase
 
     %#ok<*NASGU,*ASGLU> unused variables
 
-     properties(TestParameter)
+    properties(TestParameter)
         prec = {'double', 'single'}%, 'halfT'}
         complexity = struct('real', 'real', 'complex', 'complex');
         dev = getdevs()
-     end
+    end
 
-     properties
-         dargs % data arguments
-     end
+    properties
+        dargs % data arguments
+    end
+    methods
+        function setDev(test, dev)
+            import matlab.unittest.fixtures.TemporaryFolderFixture;
+            import matlab.unittest.fixtures.CurrentFolderFixture;
+
+            % Create a temporary folder and make it the current working folder.
+            tempFolder = test.applyFixture(TemporaryFolderFixture);
+            test.applyFixture(CurrentFolderFixture(tempFolder.Folder));
+
+            % The test can now write files to the current working folder.
+            switch class(dev)
+                case "double",                  dtype = "none";
+                case "parallel.gpu.CUDADevice", dtype = "gpu";  gpuDevice(dev.Index);
+                case "oclDevice",               dtype = "ocl";  oclDevice(dev.Index);
+            end
+            odevs = setdiff(["gpu", "ocl"], dtype);
+            for odev = odevs
+                localwritelines("function n = "+odev+"DeviceCount(), n = 0;", odev+"DeviceCount.m");
+                test.addTeardown(@()delete(fullfile(pwd, odev+"DeviceCount.m")));
+            end
+        end
+    end
 
     methods(TestClassSetup)
         % Setup for the entire class
@@ -41,6 +63,8 @@ classdef KernTest < matlab.unittest.TestCase
             test.dargs{3} = {us, b};
         end
     end
+
+    methods function set_device(test, dev), test.setDev(dev); end end
 
     % dispatch
     % Github test routine
@@ -78,6 +102,11 @@ classdef KernTest < matlab.unittest.TestCase
             import matlab.unittest.constraints.RelativeTolerance;
             if prec == "halfT", test.assumeTrue(logical(exist('halfT', 'class'))); end
 
+            % device
+            test.set_device(dev);
+            isgpu = isa(dev, 'parallel.gpu.CUDADevice');
+            isocl = isa(dev, 'oclDevice');
+
             % random values
             [A, B, N, M] = deal(test.dargs{2}{:});
             [A,B] = dealfun(str2func(prec), A, B); % cast to type
@@ -87,7 +116,7 @@ classdef KernTest < matlab.unittest.TestCase
                 case "halfT" , tol = 1e2;
             end
             switch complexity, case "real", [A,B] = deal(real(A), real(B)); end
-            switch dev, case "gpu", [A,B] = dealfun(@gpuArray, A,B); end
+            if isgpu, [A,B] = dealfun(@gpuArray, A,B); end
 
             % check each type of argument
             % TODO: move this to setup params
@@ -105,7 +134,7 @@ classdef KernTest < matlab.unittest.TestCase
             % convolve and check outputs
             for s = shapes, for d = dims %#ok<ALIGN> 
                 z1 = cell2mat(cellfun(@(A,B) {gather(conv(A, B, char(s)))}, num2cell(A,1), num2cell(B,1)));
-                z2 = gather(convd(shiftdim(A,1-d),shiftdim(B,1-d),d,s,'gpu',dev == "gpu", 'ocl', dev == "ocl"));
[…]
