### 2024-07-17 a32ab10
das_spec : wrap with try-catch
