### 2024-07-17 e956c81
testing.yml : add logging on dispatch w/ tmate.
