+++ .github/workflows/testing.yml
@@ -37,6 +37,7 @@ jobs:
           command: |
             clogdir = fullfile(pwd, 'logs','client');
             wlogdir = fullfile(pwd, 'logs','workers');
+            mkdir(clogdir);
             parallel.internal.logging.enableClientLogging(clogdir, 6)
             setSchedulerMessageHandler(@disp);
             setenv('MDCE_DEBUG','true');
@@ -48,10 +49,11 @@ jobs:
             rep = arrayfun(@(r) r.Details.DiagnosticRecord.Report, res(i), 'UniformOutput', false);
             rep{:}
 
-            c = parcluster('local')
+            c = parcluster('local'),
             fld = c.JobStorageLocation,
             fls = dir(fullfile(fld,'**','*.log*'));
             fls = string({fls.folder}, {fls.name});
+            mkdir(wlogdir);
             arrayfun(@(s) copyfile(s, wlogdir), fls);
 
       - name: Upload Logs
