### 2024-07-17 2808447
Take 2.
