### 2024-07-22 f1ef6ab
setup() : blacklist msfm Mex files; restrict 'exist' to files.
