### 2024-07-23 70bc02c
Dev (#36)
* Adding Waveform.reverse(); update documentation.

* ChannelData : add convt(), uminus(), minus()

* mod2db() : cast integer inputs to double.

* Waveform : add conj(); update documentation for conj(), reverse()

* ChannelData : added conj, uplus, rdivide, ldivide; ChdTest : update arithmetic tests.

* ChannelData.scale : accept array inputs.

* pwznxcorr : adding (experimental) upsampling option.

* UltrasoundSystem.focusTx() : set default block size to all Txs; cosmetic

* wsinterpd2() : CUDA/OpenCL kernel calls cached; avoid OpenCL file lookup if native CUDA branch selected.

* d
[…]
