### 2024-07-24 #6688 @adamltyson
Thanks @tomelse! I'll add you to the review over at https://github.com/openjournals/joss-reviews/issues/6772.

@ebezzam, if you'd still like to review this too, let me know. 3 sets of eyes is better than 2!
