### 2024-07-26 9cfcdf0
Update debug.yml
