### 2024-07-26 3b3e06a
Update debug.yml
