### 2024-07-26 14f4ac8
Update debug.yml
