### 2024-07-28 ed6209c
testing.yml : re-run failed tests once in case of random failures.
