+++ .github/workflows/testing.yml
@@ -26,8 +26,12 @@ jobs:
         with:
           command: |
             setup cache; addpath build;
-            res = runProjectTests("verbosity", "Verbose"), 
+            [res, suite, runner] = runProjectTests("verbosity", "Verbose"),
             i = find([res.Failed]); res(i).Name, 
+            rep = arrayfun(@(r) r.Details.DiagnosticRecord.Report, res(i), 'UniformOutput', false);
+            rep{:}
+            disp("Attempting retry of "+nnz(i)+" failed jobs.");
+            res(i) = runner.run(suite(i));
             rep = arrayfun(@(r) r.Details.DiagnosticRecord.Report, res(i), 'UniformOutput', false);
             rep{:}
             assertSuccess(res);
