### 2024-07-28 72a008b
ChannelData.Verasonics() : added support for custom sampling configurations; all modes return at decimSampleRate.
