### 2024-08-05 #6772 @tomelse
@thorstone25 Would you mind explaining a bit why under installation "failures can be safely ignored"? My installation gave an error: "Error using buildtool (line 50). Build failed." Then gave an error when I tried to run the tests ("Unrecognised function or variable 'gpuDeviceCount'").
