### 2024-08-05 thorstone25/qups#37 @tomelse
Very strange bug
Hi :) I'm the reviewer for JOSS ([#6772](https://github.com/openjournals/joss-reviews/issues/6772)).

I think I've found a slightly bizarre bug... On my linux system, my home directory is the following (anonymised): `/home/xyz.cxyz.org/name`. When trying to run your tests, it means that I get the following error:

```
** Starting compile_mex_msfm2d
## -----------------------------------------------------------------------------
## Error occurred in 'compile_mex_msfm2d' and it did not run to completion.
##   Identifier: 'MATLAB:mex:AccessDenied'
##      Message: Access denied; check that you have permissions to access '/home/xyz.mexa64xyz.org/name/Documents/qups/bin'.
##        Stack: /usr/local/MATLAB/R2024a/toolbox/matlab/buildtool/ext/+matlab/+buildtool/+tasks/MexTask.m (MexTask.buildMex) at 172
## -----------------------------------------------------------------------------
** Failed compile_mex_msfm2d
```

I've tracked it down to this line, I think: `ofls = replace(fullfile(base, "bin", nms), '.c', "."+mexext());`. Clearly you were expecting the path to end in .c, I guess, but by some quirk it is in my home directory path.
