### 2024-08-05 thorstone25/qups#37 @tomelse
Okay I've managed to hack it to work on my local PC, replacing the following line (line 84 in buildfile.m).

```
oils = replace(fullfile(base, "bin", nms), ".c", "."+mexext());
```

I replaced with:

```
for i = 1:numel(nms)
[~, stem] = fileparts(nms(l));
ofls(i) = fullfile(base, "bin", stem+"."+mexext());
end
```
