### 2024-08-06 thorstone25/qups#37 @thorstone25
Hmm, very strange indeed -  I'm noticing though that the erroring path starts with
`/home/xyz**.mexa64**xyz.org ...`
- I'm guessing some folder in the path matched the `".c"` token.
This looks like a bit of a corner case, but a bug for sure, glad you caught it!
