### 2024-08-06 #6772 @thorstone25
> @thorstone25 I notice you're the sole author. Could you confirm that the contributions of others was not sufficient to warrant co-authorship? Thanks!

Hi @tomelse, yes I can confirm sole authorship. This work doesn't really involve many novel concepts or ideas, it's mainly an optimized implementation and convention library. As for the other contributors, only #23 added any functionality - the rest were template files, typos, or removed.
