### 2024-08-06 #6772 @thorstone25
> > @thorstone25 Would you mind explaining a bit why under installation "failures can be safely ignored"? My installation gave an error: "Error using buildtool (line 50). Build failed." Then gave an error when I tried to run the tests ("Unrecognised function or variable 'gpuDeviceCount'").
>
> I think you need to include the parallel computing toolbox as a dependency? I'm not an expert in MATLAB, so I'm not really sure how you might include this in your installation? EDIT: possibly also the signal processing toolbox?
>
> I'm also working on an issue on the main repository, so please keep an eye out and try and address ASAP please. Thanks :)

Yes, the signal processing toolbox and parallel computing toolboxes are the only 2 dependencies! I missed that, thank you - I'll update the README.
