### 2024-08-06 #6772 @thorstone25
> @thorstone25 Would you mind explaining a bit why under installation "failures can be safely ignored"? My installation gave an error: "Error using buildtool (line 50). Build failed." Then gave an error when I tried to run the tests ("Unrecognised function or variable 'gpuDeviceCount'").

Well, `gpuDeviceCount()` is part of the parallel computing toolbox, so actually this particular failure is pointing to a larger dependency issue as you noted already, but in general, cloning the repo and opening the project is sufficient to use QUPS. These failures are the failure to install optional functions that don't affect core functionality.

Even the mex task failing in thorstone25/qups#37 can be safely ignored because a native m-file implementation is available. The most common failure will be for opening the Matlab-OpenCL project, which attempts to compile OpenCL Mex files assuming a working installation of OpenCL, which most users (myself included) won't install, and won't really need - I expect pretty much all users to see that error.
