### 2024-08-06 78297ea
Bug fix for #37
