### 2024-08-06 thorstone25/qups#37 @tomelse
Thanks! Definitely a weird edge case, but I think a good reminder that we should use specialised file path libraries rather than string operations for these sorts of things
