### 2024-08-06 #6772 @tomelse
Thanks for clarifying! That all sounds good to me. Maybe it's worth clarifying all of this a little bit in the README, so if people see a genuine error, they can debug it a bit more easily. This is only a suggestion, by no means a requirement for the review.
