### 2024-08-06 e985700
buildfile : add 'compatability' task to report installed toolboxes vs. requirements.
