+++ buildfile.m
@@ -49,7 +49,6 @@ plan("patch") = matlab.buildtool.Task( ...
     "Dependencies", "patch_"+[ext_nms([2 4 3])] ...
     );
 
-
 %% Testing and Artifacts
 % source files (folders)
 src = ["kern","src","utils"] + filesep;
@@ -72,6 +71,7 @@ plan("test") = TestTask( ...
     "Tag","syntax","SourceFiles",src ...
     ,CodeCoverageResults=fullfile("build","code-coverage-brief",["coverage.xml" "html/index.html"]) ...
     ,TestResults="build/test-results-brief/report.html" ...
+    ,Dependencies="compatability"...
     );
 
 %% Add compilation dependencies
@@ -415,5 +415,20 @@ ofl  = context.Task.Outputs.Path;
 res = runProjectTests("benchmark", 0);
 txt = string({cat(2,cat(2,res.Details).DiagnosticRecord).Report})';
 writelines(txt, ofl);
+
+end
+
+function compatabilityTask(context)
+% Check for required toolboxes
+req = ["Signal Processing", "Parallel Computing"];
+rec = ["Image Processing"];
+sug = reshape(string.empty,1,0);
+rep = "The following packages are " + ["required", "recommended", "suggested"] ... 
+    + ": " + sprintf('\t') + cellfun(@(s)join(s,", ",2),{req,rec,sug});
+arrayfun(@disp, rep(~ismissing(rep)));
+vn = string({ver().Name});
+i = ismember(req + " Toolbox", vn);
+assert(all(i), "The following toolboxes are required, but not installed:"+newline+join(req(~i)+ " Toolbox", newline));
+disp("All required toolboxes are installed.");
 
 end
