### 2024-08-06 thorstone25/qups#39 @tomelse
Warnings when running tests
![Screenshot 2024-08-06 at 11 01 06](https://github.com/user-attachments/assets/c4d72133-fa9d-4ec6-b24f-3a4376ecfbc7)

I'm getting several warnings (which will become errors in future releases) when running tests. I think this is something you should address, as it will ensure the long-term stability of your package. (Review for JOSS [#6772](https://github.com/openjournals/joss-reviews/issues/6772)).

Additionally, the tests are producing a huge amount of output, I don't really know how MATLAB tests are structured, so I don't know how easy it is to reduce this. I think it would be really helpful to reduce the amount of text produced to make it more interpretable.
