### 2024-08-06 thorstone25/qups#39 @tomelse
There's several other deprecation warnings that I've spotted. Too many to include in here. I think it's important to try and address as many of these as possible to ensure that the package has some longevity. Take a look at the tests on a fresh install and see if you can fix a few of them. Thanks
