### 2024-08-06 1cafe53
greens() : bug fix (addresses #39)
