### 2024-08-07 b5c3afe
Suppress test outputs - addresses #39
