+++ test/ExampleTest.m
@@ -7,6 +7,9 @@ classdef ExampleTest < matlab.unittest.TestCase
         scat_lim (1,1) double % limit of scatterers
         pulse_lim (1,2) double % max number of pulses to sim in k-Wave, greens
     end
+    properties(Constant)
+        debug (1,1) logical = false; % whether to print debug warnings
+    end
     
     % blacklists
     properties
@@ -56,9 +59,9 @@ classdef ExampleTest < matlab.unittest.TestCase
                 'FieldII',  ["calc_scat"+["","_all", "_multi"], "getFieldII"+["Aperture","Patches","Positions"]], ... FieldII
                 'MUST',     "simus", ... MUST
                 'fullwave', ["getFullwaveTransducer", "fullwaveSim", "fullwaveConf", "fullwaveJob", "mapToCoords"], ... % fullwave
-                'mex',      ["bfEikonal", "msfm"+["","2d","3d"]], ... mex binaries required ... compilation (optional, requires CUDA)
+                'mex',      "recompile" + ["","Mex"], ... mex binaries required ... compilation (optional, requires CUDA)
                 'gpu',      ["wbilerpg","feval"], ... CUDAKernel or oclKernel support required
-                'comp',     ("recompile" + ["","Mex","CUDA"]), ... compilations setup required
+                'comp',     ("recompile" + ["","CUDA"]), ... compilations setup required
                 'ext',      "vol3d", ... external functions
                 }, 2, [])'; % + "(";
             bl_fcn(:,2) = cellfun(@(s){s+"("}, bl_fcn(:,2)); % append "("
@@ -93,12 +96,14 @@ classdef ExampleTest < matlab.unittest.TestCase
             if all(arrayfun(@(k) exist(fullfile(prj_rt, "bin", k), 'file'), kerns))
                 bl_fcn(bl_fcn(:,1) == "gpu",:) = []; 
             end
-            kerns = ["msfm2d", "msfm3d"]+"."+mexext; % require mex binaries
-            if 1||all(arrayfun(@(k) exist(fullfile(prj_rt, "bin", k), 'file'), kerns))
+            kerns = ["msfm2d", "msfm3d"]+".c"; % require mex source files
+            lst = arrayfun(@(k) dir(fullfile(prj_rt, "**", k)), kerns, 'uni', 0);
+            if all(ismember(kerns, {vertcat(lst{:}).name}))
+            % if all(arrayfun(@(k) exist(fullfile(prj_rt, "bin", k), 'file'), kerns))
                 bl_fcn(bl_fcn(:,1) == "mex",:) = []; 
             end
             % fullwave executable must exist in 'bin' folder under this name
-            kerns = ["fullwave2_executable"];
+            kerns = "fullwave2_executable";
             if all(arrayfun(@(k) exist(fullfile(prj_rt, "bin", k), 'file'), kerns)) ...
                     && exist("mapToCoords", "file")
                 bl_fcn(bl_fcn(:,1) == "fullwave",:) = []; 
@@ -136,22 +141,22 @@ classdef ExampleTest < matlab.unittest.TestCase
                 ,"Matching packages: " + join(string(pkg(1,p)), ", ") ...
                 ]), newline));
         end
-        function silenceAcceptableWarnings(testCase)
+        function silenceAcceptableWarnings(test)
             lids = [...
                 "QUPS:kspaceFirstOrder:upsampling", ...  % in US
                 "MATLAB:ver:ProductNameDeprecated", ... % in k-Wave
                 "QUPS:recompile:UnableToRecompile" ... % in US 
                 ];
                 W = warning(); % get current state
-                arrayfun(@(l)                warning(    'off', l), lids); % silence
-                testCase.addTeardown(@() warning(W)); % restore on exit
+                arrayfun(@(l) warning('off', l), lids); % silence
[…]
