+++ buildfile.m
@@ -71,7 +71,8 @@ plan("test") = TestTask( ...
     "Tag","syntax","SourceFiles",src ...
     ,CodeCoverageResults=fullfile("build","code-coverage-brief",["coverage.xml" "html/index.html"]) ...
     ,TestResults="build/test-results-brief/report.html" ...
-    ,Dependencies="compatability"...
+    ,Dependencies="compatability" ...
+    ,OutputDetail="Concise" ...
     );
 
 %% Add compilation dependencies
