### 2024-08-07 thorstone25/qups#39 @thorstone25
> ![Screenshot 2024-08-06 at 11 01 06](https://private-user-images.githubusercontent.com/56722586/355405522-c4d72133-fa9d-4ec6-b24f-3a4376ecfbc7.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjMwMTM4MjMsIm5iZiI6MTcyMzAxMzUyMywicGF0aCI6Ii81NjcyMjU4Ni8zNTU0MDU1MjItYzRkNzIxMzMtZmE5ZC00ZWM2LWIyNGYtM2E0Mzc2ZWNmYmM3LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA4MDclMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwODA3VDA2NTIwM1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTQ5MjJjZWQ1NTAyMGU1OWUyMjc5Y2ViNTk0MTQ1MzVjOTlhMDFhM2ExZjllMGQ3YmUxYWNmNjJlZDI5ZDU0NzcmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.HhbSV0qCVh1huT-bcdGm8SZuIeXaJTke8HnfL0n_HHk)
>
> I'm getting several warnings (which will become errors in future releases) when running tests. I think this is something you should address, as it will ensure the long-term stability of your package. (Review for JOSS [#6772](https://github.com/openjournals/joss-reviews/issues/6772)).
>
> Additionally, the tests are producing a huge amount of output, I don't really know how MATLAB tests are structured, so I don't know how easy it is to reduce this. I think it would be really helpful to reduce the amount of text produced to make it more interpretable.

Thanks for the feedback! I've added a bug fix for that deprecation warning and some logic to suppress most of those text outputs.

I think most of the other warnings were generated by QUPS - I'm marking properties for deprecation for a (potential) v2.x because I'm attempting to support partial backwards compatibility across v1.x for a little while, which is critical for loading mat-files saved with an older version of the package.
