### 2024-08-07 #6772 @thorstone25
Of course, I'm more than willing to update the README! Can you help me understand though what all should be further clarified? Do you mean identifying which warnings/failures/errors are safe to ignore versus those that are not?

With the most recent changes addressing thorstone25/qups#39, all but 1 of the expected warnings should be suppressed or caught, and any tests that are unsupported by the installation if e.g. OpenCL fails to install should be filtered out, so everything else should be suspect.
