### 2024-08-07 #6772 @tomelse
Thanks! I mean any steps needed to get the tests to run on a particular machine really. So in my case, it's deleting/renaming the mex files that gave the segmentation fault on Linux. I'm just thinking from the point of view of someone who has never used the toolkit before, if they come across any barriers, it needs to be quite clear how to fix them for them to carry on trying.
