### 2024-08-08 #6772 @thorstone25
Gotcha! Perhaps what's missing is a "Troubleshooting" section?

I've built in some amount of redundancy and fail-safes when the accelerated routines (CUDA/OpenCL/mex) fail for whatever reason, but I haven't described that explicitly and how to select and manage usage when the user gets e.g. GPU-OOM errors or the mex-files cause a crash, I can definitely add that.

I do want to understand the line for responsibility between this package versus MATLAB. For example, when a linux kernel change caused [errors with copyfile](https://www.mathworks.com/matlabcentral/answers/1798895-invalid-cross-device-link-18-when-saving-a-file-on-arch-linux), it affected QUPS binaries and cluster batching routines, but the fix is ultimately with MATLAB, users can only really circumvent it.

This issue with the mex files for example is similar; that file has been working for a decade, so for it to fail now I suspect is actually due to linux/MATLAB changes. I'm leaving it up in the hope that it should resolve itself in a future MATLAB update.
