### 2024-08-08 cedddce
Update README.md
1st draft - new troubleshooting section
