### 2024-08-14 5b5d028
swapdim() : slight compute performance optimization.
