### 2024-08-15 thorstone25/qups#40 @tomelse
Beamforming Walkthrough
The first two lines of `beamforming_walkthrough.mlx` do not work in the way that I would expect.

I'm not an expert in MATLAB, but I would expect to be able to run the walkthroughs without starting a new project.

The function `openProject(which("Qups.prj"))` doesn't seem to return anything either, it seems like Qups.prj doesn't exist on my computer, or at least isn't on the path?

[JOSS Review #6772](https://github.com/openjournals/joss-reviews/issues/6772)
