### 2024-08-15 thorstone25/qups#40 @tomelse
There's also some placeholders for the linear acoustic equation etc.
