### 2024-08-15 thorstone25/qups#40 @tomelse
The verasonics data is also not included in the import example. This isn't a problem given the data file is probably too big to store on GitHub. It would be nice if you could provide an example dataset that people could use here though.
