### 2024-08-15 dab4e6b
das_spec() : add error reporting.
