### 2024-08-15 thorstone25/qups#40 @thorstone25
> The first two lines of `beamforming_walkthrough.mlx` do not work in the way that I would expect.
>
> I'm not an expert in MATLAB, but I would expect to be able to run the walkthroughs without starting a new project.
>
> The function `openProject(which("Qups.prj"))` doesn't seem to return anything either, it seems like Qups.prj doesn't exist on my computer, or at least isn't on the path?
>
> [JOSS Review #6772](https://github.com/openjournals/joss-reviews/issues/6772)

The `Qups.prj` file is in the base directory of the repo. I'm not sure why it wasn't found by the `which` command if that directory was on the path? If you ran it from another folder though, then perhaps it wasn't on the path? Would it be clearer if that line was removed?
