### 2024-08-15 thorstone25/qups#40 @tomelse
I realised the problem actually is if you don't have a project open, the `~=` comparison does not work as intended. Sorry I can't provide the exact code right now as I'm not on my laptop.
