+++ paper.md
@@ -33,7 +33,7 @@ The lack of data and terminology standardization in ultrasound technology resear
 At its lowest level, `QUPS` provides memory efficient definition classes capable of representing a wide variety of common imaging systems and enforces standardized units, coordinate systems, and nomenclature. 
 All classes have data visualization support, making it far easier to share and validate data amongst collaborators and with the public. 
 
-`QUPS` provides a variety of compute and memory optimized implementations of common signal processing methods (e.g. beamforming, resampling, or convolution), which take advantage of native parallelization tools within MATLAB or hardware acceleration via CUDA or OpenCL when feasible by leveraging `MatCL`[@heinisch2018MatCL]. `QUPS` exploits data parallellism primarily using a single program multiple data (SPMD) approach, and optimistically dispatches to the most efficient implementation available on a user's computer. For access to cutting edge research, it can also export or import to `USTB`[@USTB2017], an ultrasound toolbox which contains a library of advanced algorithms from the literature.
+`QUPS` provides a variety of compute and memory optimized implementations of common signal processing methods (e.g. beamforming, resampling, or convolution), which take advantage of native parallelization tools within MATLAB or hardware acceleration via CUDA or OpenCL when feasible by leveraging `MatCL`[@heinisch2018MatCL]. `QUPS` exploits data parallelism primarily using a single program multiple data (SPMD) approach, and optimistically dispatches to the most efficient implementation available on a user's computer. For access to cutting edge research, it can also export or import to `USTB`[@USTB2017], an ultrasound toolbox which contains a library of advanced algorithms from the literature.
 
 To facilitate data acquisition, `QUPS` provides tools to import data from [Verasonics Vantage](https://verasonics.com/vantage-systems/) ultrasound research scanners, which uses a flexible yet highly complex data standard. Similarly, to facilitate data generation, `QUPS` also interfaces with multiple popular free ultrasound field simulators including `FieldII`[@Jensen1996], `MUST`[@Garcia2021], and `k-Wave`[@Treeby2010]. Simulations can be batched on a compute cluster via MATLAB's parallel server toolbox.
 
@@ -42,7 +42,7 @@ Alternatively, when a coarse but fast approximation is preferable, `QUPS` offers
 $$x_{mn}(t) = \sum_s \frac{v(t - c_0^{-1} \left( ||\mathbf{p}_s - \mathbf{q}_m|| + ||\mathbf{p}_s - \mathbf{r}_n|| \right) )}{||\mathbf{p}_s - \mathbf{q}_m|| \cdot ||\mathbf{p}_s - \mathbf{r}_n||}$$
 
 where $t$ is time, $c_0$ is the speed of sound, $v(\cdot)$ is the transmit waveform, $m$, $n$, and $s$ are the transmit, receive, and scatterer indices, $\mathbf{q}_m$ and $\mathbf{r}_n$ and $\mathbf{p}_s$ are the transmit elements, receive elements, and scatterer locations in 3D Cartesian coordinates, and $x_{mn}(\cdot)$ is the received echo waveform as a function of time.
-This expression is embarassingly parallel over $m$ and $n$, and can be further accelerated with a branch-and-bound approach used to minimize unnecessary computation. Both of these aspects are exploited in native MATLAB, as well as in CUDA and OpenCL kernels.
+This expression is embarrassingly parallel over $m$ and $n$, and can be further accelerated with a branch-and-bound approach used to minimize unnecessary computation. Both of these aspects are exploited in native MATLAB, as well as in CUDA and OpenCL kernels.
 
 Broadcasting conventions are employed throughout to reduce the memory footprint while offering significant flexibility. 
 For example, to implement delay-and-sum beamforming, one of the most common operations in ultrasound imaging, `QUPS` provides a GPU-enabled method employing the equation
[…]
