### 2024-08-20 #6772 @adamltyson
Thanks @tomelse! Thanks again for stepping in.

@thorstone25 now @tomelse and @rafaelorozco have completed their reviews, we just need to do some admin before we move to acceptance.

I've made some suggestions to the paper. Would you also be able to double check that there's no DOI for "FIELD: A program for simulating ultrasound systems"? I couldn't find one myself.

After that, would you be able create a release of the software. and archive the release on Zenodo/figshare/etc and post the version number and DOI here?

In the next comment I will bring up a checklist of some other checks too. Could you take a look and tell me when you think they're all ok?
