### 2024-08-21 79ea6b4
interpd.cu : kernel optimizations
