### 2024-08-22 fb71566
Fix typos (#42)
Co-authored-by: thorstone25 <thorstone25@users.noreply.github.com>
