### 2024-08-22 fa346dd
Update CITATION.cff, build/coverage.xml
