+++ CITATION.cff
@@ -6,7 +6,7 @@ authors:
   orcid: "https://orcid.org/0000-0002-0720-6955"
 title: "Quick Ultrasound Processing & Simulation"
 type: software
-version: v1.1.1
+version: v1.2.1
 license: Apache-2.0
-date-released: 2024-04-07
+date-released: 2024-08-22
 repository-code: "https://github.com/thorstone25/qups"
