### 2024-08-23 #6772 @thorstone25
> Thanks @tomelse! Thanks again for stepping in.
>
> @thorstone25 now @tomelse and @rafaelorozco have completed their reviews, we just need to do some admin before we move to acceptance.
>
> I've made some suggestions to the paper. Would you also be able to double check that there's no DOI for "FIELD: A program for simulating ultrasound systems"? I couldn't find one myself.
>
> After that, would you be able create a release of the software. and archive the release on Zenodo/figshare/etc and post the version number and DOI here?
>
> In the next comment I will bring up a checklist of some other checks too. Could you take a look and tell me when you think they're all ok?

@adamltyson, thanks for the quick turn-around!

I can confirm that there's no DOI for FieldII - I grabbed the most relevant citation listed on the [website](https://www.field-ii.dk/?background.html), which seems to be from a relatively small regional conference.

I've made a new release incorporating all reveiwer edits and set things up with zenodo. The concept DOI is 10.5281/zenodo.13363956 or the version v1.2.1 DOI is 10.5281/zenodo.13363957. I can also verify that the version, title, author, ORCID, and license are all correct on zenodo, and the paper.
