### 2024-08-27 #6772 @adamltyson
Thanks @thorstone25! Would you mind changing the zenodo archive title so that it matches the paper?
