### 2024-08-28 a792818
Adding private num2cell.
