### 2024-08-29 #6772 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
