### 2024-09-05 #6772 @Kevin-Mattheus-Moerman
@thorstone25 as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Everything seems to be order, so I will now proceed to accept this submission.
