### 2024-09-06 #6772 @thorstone25
@Kevin-Mattheus-Moerman thank you for confirming!

> :warning: Couldn't acccept/publish paper. [An error happened](https://github.com/openjournals/joss-papers/actions/runs/10717424954?check_suite_focus=true).

I just took a brief look at the logs here, and it looks to be a "null index" error - I'm guessing it could have something to do with the "Field II" reference not having a DOI (none exists). Is there some special formatting that needs to be done there?
