### 2024-09-11 4d8c987
ChannelData.Verasonics() : bug fix.
