+++ src/ChannelData.m
@@ -297,7 +297,7 @@ classdef ChannelData < matlab.mixin.Copyable
                 % get relevant receive info
                 b = kwargs.buffer(i);
                 Rx = Receive((b == [Receive.bufnum])); % filter by buffer and frame
-                if any(kwargs.frames), Rx = Rx(ismember([Receive.framenum], kwargs.frames)); end
+                if any(kwargs.frames), Rx = Rx(ismember([Rx.framenum], kwargs.frames)); end
                 if isempty(Rx) 
                     warning("No data found for buffer " + kwargs.buffer(i) + "."); 
                     [smode(i), fmod(i), chd(i)] = deal("N/A", nan, ChannelData('order','TMNF'));
