### 2024-09-12 #6772 @Kevin-Mattheus-Moerman
@openjournals/dev can you help address this error? :point_up:
